requirements.txt
alabaster==0.7.10
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.10
awesome-slugify==1.6.5
Babel==2.4.0
bcrypt==3.1.7
beautifulsoup4==4.5.3
boto==2.46.1
bsdiff4==1.1.9
certifi==2020.12.5
cffi==1.13.2
chardet==4.0.0
conda==4.3.16
coverage==4.3.4
cryptography==3.3.2
decorator==4.0.11
defusedxml==0.5.0
distlib==0.3.2
dj-database-url==0.4.2
Django==1.11.29
django-allauth==0.31.0
django-anymail==0.8
django-braces==1.11.0
django-ckeditor==6.1.0
django-cors-headers==3.8.0
django-coverage-plugin==1.5.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-environ==0.4.1
django-extensions==1.7.7
django-hvad==1.7.0
django-js-asset==1.2.2
django-libs==1.67.12
django-model-utils==2.6.1
django-photologue==3.7
django-redis==4.7.0
django-sortedm2m==1.4.0
django-storages-redux==1.3.2
django-taggit==0.22.2
django-test-plus==1.0.17
docutils==0.13.1
dropbox==8.3.1
environ==1.0
example==0.1.0
ExifRead==2.1.2
factory-boy==2.8.1
Faker==0.7.10
filelock==3.0.12
flake8==3.3.0
gevent==21.8.0
greenlet==1.1.1
gunicorn==19.7.1
heroku==0.1.4
httplib2==0.10.3
idna==2.10
imagesize==0.7.1
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.2.0
Jinja2==2.9.5
jmespath==0.9.2
MarkupSafe==2.0.1
mccabe==0.6.1
mysqlclient==2.0.3
node==0.9.18.1
npm==0.1.1
oauthlib==2.0.2
odict==1.5.2
olefile==0.44
optional-django==0.1.0
paramiko==2.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==8.3.1
plumber==1.3.1
prompt-toolkit==1.0.14
protobuf==3.17.3
psycopg2-binary==2.8.6
ptyprocess==0.5.1
py==1.4.33
py-moneyed==0.6.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.17
pyflakes==1.5.0
Pygments==2.2.0
PyNaCl==1.3.0
pyparsing==2.2.0
pytest==3.0.7
pytest-django==3.1.2
python-dotenv
pytest-sugar==0.8.0
python-dateutil==1.5
python3-openid==3.1.0
pytz==2017.2
pyusb==1.0.2
redis==2.10.5
regex==2017.4.5
requests==2.13.0
requests-oauthlib==0.8.0
rsa==3.4.2
ruamel.yaml==0.17.14
ruamel.yaml.clib==0.2.6
scp==0.13.2
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.3
sqlparse==0.2.3
termcolor==1.1.0
traitlets==4.3.2
Unidecode==0.4.20
uritemplate==3.0.0
urllib3==1.20
virtualenv==20.4.7
wcwidth==0.1.7
Werkzeug==0.11.15
whitenoise==3.3.0
zope.component==4.3.0
zope.deprecation==4.2.0
zope.event==4.2.0
zope.interface==5.4.0
zope.lifecycleevent==4.1.0
